feat: package folder restrucuring
