+++ lib/config/config.dart
@@ -1,1 +1,3 @@
+export 'image_margins.dart';
 export 'webview_configuration.dart';
+export 'dimension_strategy.dart';
