+++ lib/config/dimension_strategy.dart
@@ -1,34 +1,3 @@
-/// Represents the margins for an output image.
-class ImageMargins {
-  final int top;
-  final int right;
-  final int bottom;
-  final int left;
-
-  const ImageMargins({
-    this.top = 0,
-    this.right = 0,
-    this.bottom = 0,
-    this.left = 0,
-  }) : assert(top >= 0 && right >= 0 && bottom >= 0 && left >= 0);
-
-  const ImageMargins.symmetric({
-    int horizontal = 0,
-    int vertical = 0,
-  })  : assert(horizontal >= 0 && vertical >= 0),
-        left = horizontal,
-        right = horizontal,
-        top = vertical,
-        bottom = vertical;
-
-  const ImageMargins.all(int value)
-      : assert(value >= 0),
-        left = value,
-        right = value,
-        top = value,
-        bottom = value;
-}
-
 /// Utility class providing JavaScript snippets for retrieving HTML content dimensions.
 class HtmlDimensionStrategy {
   final int? width;
@@ -39,18 +8,22 @@ class HtmlDimensionStrategy {
       : width = null,
         height = null,
         script = null;
+
   const HtmlDimensionStrategy.withDimensions({
     this.width,
     this.height,
   }) : script = null;
+
   const HtmlDimensionStrategy.fitContent()
       : width = null,
         height = null,
         script = fitContentJs;
+
   const HtmlDimensionStrategy.fullScroll()
       : width = null,
         height = null,
         script = fullScrollJs;
+
   const HtmlDimensionStrategy.customScript(this.script)
       : width = null,
         height = null;
