+++ lib/html_to_image_platform_interface.dart
@@ -1,6 +1,5 @@
 import 'dart:typed_data';
 
-import 'package:html_to_image/config.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'config/config.dart';
